backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
tensorflow-cpu==2.16.1
opencv-python-headless==4.10.0.84
Pillow==10.3.0
numpy==1.26.4
libsql-client==0.3.1
boto3==1.34.131
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psutil==5.9.8
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1
